🦭 调整记忆设置页 Tab 顺序
