+++ src/i18n/locales/ar.json
@@ -971,6 +971,10 @@
     "skillHealth": "الصحة",
     "memory": "الذاكرة",
     "memoryDesc": "ذاكرات دائمة تساعد الذكاء الاصطناعي على تذكر السياق عبر المحادثات",
+    "memoryTabs": {
+      "settings": "إعدادات الذاكرة",
+      "manage": "إدارة الذاكرة"
+    },
     "activeMemoryTitle": "استدعاء الذاكرة النشطة",
     "activeMemoryDesc": "في بداية كل دور من أدوار المستخدم، نفّذ استعلامًا جانبيًا محدودًا لاستخلاص أكثر ذكرى ذات صلة وحقنها ضمن مطالبة النظام. الأخطاء والمهلات تتراجع بصمت إلى الاستدعاء السلبي ولا تعرقل المحادثة أبدًا.",
     "activeMemoryTimeout": "المهلة (مللي ثانية)",
@@ -2993,6 +2997,7 @@
     }
   },
   "planMode": {
+    "enter": "الدخول إلى Plan Mode",
     "enterDialog": {
       "header": "Plan Mode",
       "question": "هل تريد الدخول إلى Plan Mode؟ سيستكشف النموذج أولاً، ويطرح أسئلة توضيحية، ويُعدّ خطة مكتوبة لمراجعتك قبل البدء بالعمل.",
@@ -3009,7 +3014,6 @@
         }
       }
     },
-    "enter": "الدخول إلى Plan Mode",
     "toggleLabel": "خطة",
     "indicator": "وضع التخطيط",
     "plan": "خطة التنفيذ",
