+++ src/i18n/locales/es.json
@@ -971,6 +971,10 @@
     "skillHealth": "Estado",
     "memory": "Memoria",
     "memoryDesc": "Memorias persistentes que ayudan a la IA a recordar contexto entre conversaciones",
+    "memoryTabs": {
+      "settings": "Configuración de memoria",
+      "manage": "Gestión de memoria"
+    },
     "activeMemoryTitle": "Recuerdo activo de memoria",
     "activeMemoryDesc": "Al inicio de cada turno del usuario, ejecuta una side query acotada para destilar la memoria más relevante en una frase e inyectarla junto al prompt del sistema. Los tiempos de espera y fallos caen silenciosamente al recall pasivo — nunca bloquean el chat.",
     "activeMemoryTimeout": "Tiempo de espera (ms)",
@@ -2993,6 +2997,7 @@
     }
   },
   "planMode": {
+    "enter": "Entrar en Plan Mode",
     "enterDialog": {
       "header": "Plan Mode",
       "question": "¿Entrar en Plan Mode? El modelo explorará primero, hará preguntas de clarificación y redactará un plan por escrito para tu revisión antes de empezar.",
@@ -3009,7 +3014,6 @@
         }
       }
     },
-    "enter": "Entrar en Plan Mode",
     "toggleLabel": "Plan",
     "indicator": "Plan Mode",
     "plan": "Plan de ejecución",
