+++ src/i18n/locales/ko.json
@@ -971,6 +971,10 @@
     "skillHealth": "상태",
     "memory": "기억",
     "memoryDesc": "AI 가 대화 간 컨텍스트를 기억하도록 돕는 영구 기억",
+    "memoryTabs": {
+      "settings": "기억 설정",
+      "manage": "기억 관리"
+    },
     "activeMemoryTitle": "능동 메모리 리콜",
     "activeMemoryDesc": "사용자 턴이 시작될 때마다 제한된 side query를 실행하여 가장 관련된 메모리를 한 문장으로 추출해 시스템 프롬프트에 주입합니다. 타임아웃과 실패는 조용히 수동 리콜로 폴백되며 대화를 막지 않습니다.",
     "activeMemoryTimeout": "시간 초과 (ms)",
@@ -2993,6 +2997,7 @@
     }
   },
   "planMode": {
+    "enter": "계획 모드 진입",
     "enterDialog": {
       "header": "계획 모드",
       "question": "계획 모드로 들어갈까요? 모델이 먼저 탐색하고 명확화 질문을 한 뒤, 작업 전에 검토용 계획서를 작성합니다.",
@@ -3009,7 +3014,6 @@
         }
       }
     },
-    "enter": "계획 모드 진입",
     "toggleLabel": "계획",
     "indicator": "계획 모드",
     "plan": "실행 계획",
