+++ src/i18n/locales/ms.json
@@ -971,6 +971,10 @@
     "skillHealth": "Kesihatan",
     "memory": "Ingatan",
     "memoryDesc": "Ingatan kekal yang membantu AI mengingati konteks merentasi perbualan",
+    "memoryTabs": {
+      "settings": "Tetapan ingatan",
+      "manage": "Pengurusan ingatan"
+    },
     "activeMemoryTitle": "Recall memori aktif",
     "activeMemoryDesc": "Pada permulaan setiap giliran pengguna, jalankan side query bertimbang untuk merumuskan memori paling relevan dalam satu ayat dan suntikkannya bersama prompt sistem. Timeout dan kegagalan akan senyap jatuh ke recall pasif — tidak pernah menghalang sembang.",
     "activeMemoryTimeout": "Timeout (ms)",
@@ -2993,6 +2997,7 @@
     }
   },
   "planMode": {
+    "enter": "Masuk Plan Mode",
     "enterDialog": {
       "header": "Plan Mode",
       "question": "Masuk Plan Mode? Model akan meneroka dahulu, bertanya soalan penjelasan, dan menyusun pelan bertulis untuk anda semak sebelum mula bekerja.",
@@ -3009,7 +3014,6 @@
         }
       }
     },
-    "enter": "Masuk Plan Mode",
     "toggleLabel": "Pelan",
     "indicator": "Plan Mode",
     "plan": "Pelan pelaksanaan",
