+++ src/i18n/locales/pt.json
@@ -971,6 +971,10 @@
     "skillHealth": "Saúde",
     "memory": "Memória",
     "memoryDesc": "Memórias persistentes que ajudam a IA a lembrar contexto entre conversas",
+    "memoryTabs": {
+      "settings": "Configurações de memória",
+      "manage": "Gerenciamento de memória"
+    },
     "activeMemoryTitle": "Recall de memória ativo",
     "activeMemoryDesc": "No início de cada turno do usuário, execute uma side query limitada para destilar a memória mais relevante em uma frase e injetá-la junto ao prompt do sistema. Timeouts e falhas caem silenciosamente no recall passivo — nunca bloqueiam a conversa.",
     "activeMemoryTimeout": "Timeout (ms)",
@@ -2993,6 +2997,7 @@
     }
   },
   "planMode": {
+    "enter": "Entrar em Plan Mode",
     "enterDialog": {
       "header": "Plan Mode",
       "question": "Entrar em Plan Mode? O modelo vai explorar primeiro, fazer perguntas de esclarecimento e redigir um plano por escrito para você revisar antes de começar.",
@@ -3009,7 +3014,6 @@
         }
       }
     },
-    "enter": "Entrar em Plan Mode",
     "toggleLabel": "Plano",
     "indicator": "Plan Mode",
     "plan": "Plano de execução",
