+++ src/i18n/locales/ru.json
@@ -971,6 +971,10 @@
     "skillHealth": "Состояние",
     "memory": "Память",
     "memoryDesc": "Постоянная память, помогающая ИИ помнить контекст между беседами",
+    "memoryTabs": {
+      "settings": "Настройки памяти",
+      "manage": "Управление памятью"
+    },
     "activeMemoryTitle": "Активный вызов памяти",
     "activeMemoryDesc": "В начале каждого хода пользователя выполняется ограниченный side query, чтобы выделить самую релевантную память в одно предложение и внедрить её рядом с системным промптом. Таймауты и сбои молча откатываются к пассивному вызову — никогда не блокируют чат.",
     "activeMemoryTimeout": "Таймаут (мс)",
@@ -2993,6 +2997,7 @@
     }
   },
   "planMode": {
+    "enter": "Войти в Plan Mode",
     "enterDialog": {
       "header": "Plan Mode",
       "question": "Войти в Plan Mode? Модель сначала изучит контекст, задаст уточняющие вопросы и подготовит письменный план для вашего согласования, прежде чем приступить к работе.",
@@ -3009,7 +3014,6 @@
         }
       }
     },
-    "enter": "Войти в Plan Mode",
     "toggleLabel": "План",
     "indicator": "Режим планирования",
     "plan": "План выполнения",
