+++ src/i18n/locales/vi.json
@@ -971,6 +971,10 @@
     "skillHealth": "Sức khỏe",
     "memory": "Bộ nhớ",
     "memoryDesc": "Bộ nhớ liên tục giúp AI ghi nhớ ngữ cảnh qua các cuộc trò chuyện",
+    "memoryTabs": {
+      "settings": "Cài đặt bộ nhớ",
+      "manage": "Quản lý bộ nhớ"
+    },
     "activeMemoryTitle": "Truy hồi bộ nhớ chủ động",
     "activeMemoryDesc": "Đầu mỗi lượt người dùng, chạy một side query có giới hạn để cô đọng ký ức liên quan nhất thành một câu và đưa vào cùng system prompt. Hết thời gian và lỗi sẽ lặng lẽ quay lại recall bị động — không bao giờ chặn cuộc trò chuyện.",
     "activeMemoryTimeout": "Hết hạn (ms)",
@@ -2993,6 +2997,7 @@
     }
   },
   "planMode": {
+    "enter": "Vào Plan Mode",
     "enterDialog": {
       "header": "Plan Mode",
       "question": "Vào Plan Mode? Mô hình sẽ khảo sát trước, hỏi làm rõ và soạn một bản kế hoạch để bạn xem xét trước khi bắt tay vào việc.",
@@ -3009,7 +3014,6 @@
         }
       }
     },
-    "enter": "Vào Plan Mode",
     "toggleLabel": "Kế hoạch",
     "indicator": "Plan Mode",
     "plan": "Kế hoạch thực thi",
